requirements.txt
django>=6.0.0
channels>=4.0.0
daphne>=4.0.0
channels_redis>=4.2.0
Pillow>=10.0.0
djangorestframework>=3.15.0
drf-spectacular>=0.28.0
drf-spectacular-sidecar>=2024.12.1
